hello.component.js: Add a button that prints the name to the console
Clicking on the button prints the current name to the console. Normally
you wouldn't include the debug console in the controller scope, but this
is here as an example to illustrate the action. The window isn't
available to the controller's scope, it only includes what it defines on
its scope.